+++ translator/src/main/resources/se/umu/cs/rai/scopus/translator/documentation/Manual.scala
@@ -140,3 +140,13 @@ case class FactorialVerbose () extends AbstractFactorialVerbose {
     else factorial_rec (n - 1, n * product)
 
 }
+
+case class Main () {
+  def main (args: Array[String]) =
+    println ("Hello world!")
+}
+
+object EntryPoint {
+  def main(args: Array[String]): Unit = Main().main(args)
+}
+