+++ translator/src/main/scala/se/umu/cs/rai/scopus/translator/EntryPoint.scala
@@ -1,0 +1,5 @@
+package se.umu.cs.rai.scopus.translator
+
+object EntryPoint {
+  def main(args: Array[String]): Unit = Main().main(args)
+}